feat(api): add drizzle migrations and deploy UI
- Add drizzle-kit config for sqlite and generate
  migrations
- Include initial migration snapshot and journal
  cleanup
- Fix upsertScalingPolicy to store enabled as 0/1
  and handle undef
- Improve git deployment flow: validate commit
  SHAs and fetch FETCH_HEAD
- Introduce frontend deployment UI: history and
  logs
- Refactor deployments tab to use new components
- Enable API build in docker-compose for local dev
